# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Majed Sultan Al-Subaie Law Firm ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts/{id}/generate-pdf:
    post:
      operationId: generateContractPdf
      tags: [contracts]
      summary: Generate a PDF for the contract (SYSTEM_MANAGER only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractPdfResult"
        "404":
          description: Not found
        "500":
          description: PDF generation failed
...
components:
  schemas:
...
    ContractPdfResult:
      type: object
      required: [pdfUrl]
      properties:
        pdfUrl:
          type: string
...